Update es2015andjsx
Thinking the reference to `element` should be `createElement` here. I could be wrong, but figured it'd be easiest to make the change, propose it, and let you see that :)

Also, should the reference in .bablerc also have `createElement` instead of just `element`? There seems to be an inconsistency between the two, so not sure if either is wrong.